Add ImageLightbox and EmperimentEmpty components style guide
