+++ src/components/ExperimentEmpty/ExperimentEmpty.md
@@ -1,0 +1,5 @@
+**Default**
+
+```js
+<ExperimentEmpty />
+```
