+++ src/components/ImageLightbox/ImageLightbox.md
@@ -1,0 +1,5 @@
+**Default**
+
+```js
+<ImageLightbox plotUrl='https://www.cpqd.com.br/wp-content/themes/cpqd-2017/img/logotipo-cpqd.png' />
+```
